+++ util/tracing/src/lib.rs
@@ -1,20 +1,23 @@
 extern crate failure;
+extern crate humthreads;
 extern crate opentracingrust;
 extern crate opentracingrust_zipkin;
-
+extern crate reqwest;
 extern crate serde;
-#[macro_use]
 extern crate serde_derive;
 #[cfg(test)]
 extern crate serde_yaml;
 extern crate slog;
 
 extern crate replicante_util_failure;
+extern crate replicante_util_upkeep;
 
-use opentracingrust::utils::ReporterThread;
+use std::time::Duration;
+
 use opentracingrust::Tracer;
-
 use slog::Logger;
+
+use replicante_util_upkeep::Upkeep;
 
 mod backends;
 mod config;
@@ -25,21 +28,41 @@ pub use self::error::Error;
 pub use self::error::ErrorKind;
 pub use self::error::Result;
 
-/// Tracer-dependent additional returns options.
-///
-/// Returned along-side the tracer itself to pass back any thread needed my the tracer.
-pub enum TracerExtra {
-    /// The tracer has no extra returns to provide.
-    Nothing,
+/// Additional options passed to tracer configuration.
+pub struct Opts<'a> {
+    flush_timeout: Duration,
+    logger: Logger,
+    service_name: &'a str,
+    upkeep: &'a mut Upkeep,
+}
 
-    /// The tracer's `ReporterThread` that send spans to the backend.
-    ReporterThread(ReporterThread),
+impl<'a> Opts<'a> {
+    pub fn new<S>(service_name: S, logger: Logger, upkeep: &'a mut Upkeep) -> Opts<'a>
+    where
+        S: Into<&'a str>,
+    {
+        Opts {
+            flush_timeout: Duration::from_secs(1),
+            logger,
+            service_name: service_name.into(),
+            upkeep,
+        }
+    }
+
+    /// Set the muximum delay between span flushes.
+    ///
+    /// Some tracers' collectors allow this option to be set through the configuration.
+    /// In that case, the value from the user configuration overrides this option.
+    pub fn flush_timeout(mut self, timeout: Duration) -> Opts<'a> {
+        self.flush_timeout = timeout;
+        self
+    }
 }
 
 /// Creates a new tracer based on the given configuration.
-pub fn tracer(config: Config, logger: Logger) -> Result<(Tracer, TracerExtra)> {
+pub fn tracer(config: Config, opts: Opts) -> Result<Tracer> {
     match config {
         Config::Noop => self::backends::noop(),
-        Config::Zipkin(config) => self::backends::zipkin(config, logger),
+        Config::Zipkin(config) => self::backends::zipkin(config, opts),
     }
 }